+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clean up VPC peerings from a cluster VPC when is cluster deleted.
 - Clean up Application and Network loadbalancers created by Kubernetes when cluster is deleted.
 
+### Changed
+
+- Mount containerd socket instead of dockershim one to `aws-node` pods.
+
 ## [10.2.0] - 2021-02-08
 
 ### Added